reorder import of models so tests don't break(see)
For some tests we do a require of mosto.js, that in turn ends up
calling iobackend.patchBackbone().

This wouldn't be a problem if it happens at the beginning but within the
test suite models were created before and after that and it is not a
very good thing (for instance, Backbone-relational crashes with a
'RangeError: Maximum call stack size exceeded').

So, we patch backbone before all tests are run and also move the
require() statements inside the functions. That way they will be
evaluated only when they are run and not when the whole test suite
is built.